backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
sqlalchemy
aiosqlite
python-chess